+++ src/workers/work.ts
@@ -4,6 +4,7 @@ import { MarkdownItRenderer } from "../core/rendering/adapters/it-mardown";
 const renderer = new MarkdownItRenderer();
 
 self.onmessage = async function (e: MessageEvent<RenderRequest>) {
+  console.log("worker");
   const { id, document } = e.data;
   const renderedDocument = await renderer.render(document);
   self.postMessage({ id, document: renderedDocument });